pyproject.toml
[project]
name = "sari-data-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "datasketch>=1.6.5",
    "ipykernel>=6.29.5",
    "nltk>=3.9.1",
    "pandas>=2.2.3",
    "pydantic>=2.11.3",
    "pyyaml>=6.0.2",
    "ruff>=0.11.4",
    "tqdm>=4.67.1",
]
